# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trade Therapist API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type TradeWithJournalSide =
  (typeof TradeWithJournalSide)[keyof typeof TradeWithJournalSide];

export const TradeWithJournalSide = {
  long: "long",
  short: "short",
} as const;

export type JournalAnalysisEmotion =
  (typeof JournalAnalysisEmotion)[keyof typeof JournalAnalysisEmotion];

export const JournalAnalysisEmotion = {
  calm: "calm",
  fomo: "fomo",
  fear: "fear",
  greed: "greed",
  revenge: "revenge",
  hope: "hope",
  frustration: "frustration",
  confidence: "confidence",
} as const;

export interface JournalAnalysis {
  id: number;
  tradeId: number;
  transcript: string;
  emotion: JournalAnalysisEmotion;
  emotionScore: number;
  planFollowingScore: number;
  tags: string[];
  verdict: string;
  createdAt: string;
}

export interface TradeWithJournal {
  id: number;
  userId: string;
  ticker: string;
  side: TradeWithJournalSide;
  entryPrice: number;
  /** @nullable */
  exitPrice?: number | null;
  size: number;
  /** @nullable */
  pnl?: number | null;
  openedAt: string;
  /** @nullable */
  closedAt?: string | null;
  createdAt: string;
  journal?: JournalAnalysis | null;
}

export type CreateTradeBodySide =
  (typeof CreateTradeBodySide)[keyof typeof CreateTradeBodySide];

export const CreateTradeBodySide = {
  long: "long",
  short: "short",
} as const;

export interface CreateTradeBody {
  ticker: string;
  side: CreateTradeBodySide;
  entryPrice: number;
  /** @nullable */
  exitPrice?: number | null;
  size: number;
  /** @nullable */
  pnl?: number | null;
  openedAt: string;
  /** @nullable */
  closedAt?: string | null;
  transcript: string;
}

export type InsightsDataWinRateByEmotionItem = {
  emotion: string;
  winRate: number;
  tradeCount: number;
};

export type InsightsDataPnlByHourItem = {
  hour: number;
  avgPnl: number;
  tradeCount: number;
};

export type InsightsDataTriggerWordsItem = {
  word: string;
  count: number;
};

export type InsightsDataSummary = {
  totalTrades: number;
  totalPnl: number;
  winRate: number;
  mostCommonEmotion?: string;
  calmWinRate?: number;
  fomoWinRate?: number;
};

export interface InsightsData {
  winRateByEmotion: InsightsDataWinRateByEmotionItem[];
  pnlByHour: InsightsDataPnlByHourItem[];
  triggerWords: InsightsDataTriggerWordsItem[];
  summary: InsightsDataSummary;
}

export interface WeeklyReport {
  markdown: string;
  generatedAt: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};