# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trade Therapist API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List all trades with journal analysis
      responses:
        "200":
          description: List of trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradeWithJournal"
    post:
      operationId: createTrade
      tags: [trades]
      summary: Create a trade with voice transcript — runs AI analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTradeBody"
      responses:
        "201":
          description: Created trade with AI analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeWithJournal"
  /trades/{id}:
    delete:
      operationId: deleteTrade
      tags: [trades]
      summary: Delete a trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
...
  schemas:
...
    TradeWithJournal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        ticker:
          type: string
        side:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        exitPrice:
          type: [number, 'null']
        size:
          type: number
        pnl:
          type: [number, 'null']
        openedAt:
          type: string
          format: date-time
        closedAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        journal:
          oneOf:
            - $ref: '#/components/schemas/JournalAnalysis'
            - type: 'null'
      required: [id, userId, ticker, side, entryPrice, size, openedAt, createdAt]
    JournalAnalysis:
      type: object
      properties:
        id:
          type: integer
        tradeId:
          type: integer
        transcript:
          type: string
        emotion:
          type: string
          enum: [calm, fomo, fear, greed, revenge, hope, frustration, confidence]
        emotionScore:
          type: number
        planFollowingScore:
          type: number
        tags:
          type: array
          items:
            type: string
        verdict:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, tradeId, transcript, emotion, emotionScore, planFollowingScore, tags, verdict, createdAt]
    CreateTradeBody:
      type: object
      properties:
        ticker:
          type: string
        side:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        exitPrice:
          type: [number, 'null']
        size:
          type: number
        pnl:
          type: [number, 'null']
        openedAt:
          type: string
          format: date-time
        closedAt:
          type: [string, 'null']
          format: date-time
        transcript:
          type: string
      required: [ticker, side, entryPrice, size, openedAt, transcript]
...