# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trade Therapist API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get the currently authenticated user
 */
export const GetCurrentAuthUserHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const GetCurrentAuthUserResponse = zod.object({
  user: zod.union([
    zod.object({
      id: zod.string(),
      email: zod.string().email().nullable(),
      firstName: zod.string().nullable(),
      lastName: zod.string().nullable(),
      profileImageUrl: zod.string().nullable(),
    }),
    zod.null(),
  ]),
});

/**
 * @summary Start the browser OIDC login flow
 */
export const BeginBrowserLoginQueryParams = zod.object({
  returnTo: zod.coerce.string().optional(),
});

/**
 * @summary Complete the browser OIDC login flow
 */
export const HandleBrowserLoginCallbackQueryParams = zod.object({
  code: zod.coerce.string().optional(),
  state: zod.coerce.string().optional(),
  iss: zod.coerce.string().url().optional(),
});

/**
 * @summary Clear the session and begin OIDC logout
 */
export const LogoutBrowserSessionHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

/**
 * @summary Exchange a mobile OIDC code for a session token
 */

export const ExchangeMobileAuthorizationCodeBody = zod.object({
  code: zod.string().min(1),
  code_verifier: zod.string().min(1),
  redirect_uri: zod.string().url().min(1),
  state: zod.string().min(1),
  nonce: zod.string().min(1).optional(),
});

export const ExchangeMobileAuthorizationCodeResponse = zod.object({
  token: zod.string(),
});

/**
 * @summary Delete a mobile session token
 */
export const LogoutMobileSessionHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const LogoutMobileSessionResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary List all trades with journal analysis
 */
export const ListTradesResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  ticker: zod.string(),
  side: zod.enum(["long", "short"]),
  entryPrice: zod.number(),
  exitPrice: zod.number().nullish(),
  size: zod.number(),
  pnl: zod.number().nullish(),
  openedAt: zod.coerce.date(),
  closedAt: zod.coerce.date().nullish(),
  createdAt: zod.coerce.date(),
  journal: zod
    .union([
      zod.object({
        id: zod.number(),
        tradeId: zod.number(),
        transcript: zod.string(),
        emotion: zod.enum([
          "calm",
          "fomo",
          "fear",
          "greed",
          "revenge",
          "hope",
          "frustration",
          "confidence",
        ]),
        emotionScore: zod.number(),
        planFollowingScore: zod.number(),
        tags: zod.array(zod.string()),
        verdict: zod.string(),
        createdAt: zod.coerce.date(),
      }),
      zod.null(),
    ])
    .optional(),
});
export const ListTradesResponse = zod.array(ListTradesResponseItem);

/**
 * @summary Create a trade with voice transcript — runs AI analysis
 */
export const CreateTradeBody = zod.object({
  ticker: zod.string(),
  side: zod.enum(["long", "short"]),
  entryPrice: zod.number(),
  exitPrice: zod.number().nullish(),
  size: zod.number(),
  pnl: zod.number().nullish(),
  openedAt: zod.coerce.date(),
  closedAt: zod.coerce.date().nullish(),
  transcript: zod.string(),
});

/**
 * @summary Delete a trade
 */
export const DeleteTradeParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get aggregated data for charts
 */
export const GetInsightsResponse = zod.object({
  winRateByEmotion: zod.array(
    zod.object({
      emotion: zod.string(),
      winRate: zod.number(),
      tradeCount: zod.number(),
    }),
  ),
  pnlByHour: zod.array(
    zod.object({
      hour: zod.number(),
      avgPnl: zod.number(),
      tradeCount: zod.number(),
    }),
  ),
  triggerWords: zod.array(
    zod.object({
      word: zod.string(),
      count: zod.number(),
    }),
  ),
  summary: zod.object({
    totalTrades: zod.number(),
    totalPnl: zod.number(),
    winRate: zod.number(),
    mostCommonEmotion: zod.string().optional(),
    calmWinRate: zod.number().optional(),
    fomoWinRate: zod.number().optional(),
  }),
});

/**
 * @summary Generate a weekly psychology report in Dr. Trade voice
 */
export const GenerateWeeklyReportResponse = zod.object({
  markdown: zod.string(),
  generatedAt: zod.coerce.date(),
});